switched file conventions to avoid problems

+++ src/simulatorGUI/MassSpec.java
@@ -27,6 +27,8 @@ import java.io.FilenameFilter;
 import java.io.IOException;
 import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
+import java.io.UnsupportedEncodingException;
+import java.net.URLDecoder;
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
 import java.util.ArrayList;
@@ -326,9 +328,17 @@ public class MassSpec {
 		
 		// create a unique ID for this MS (used to prevent race conditions on output files)
 		msIdx = msIdxMaster++;
-		pathToClass = this.getClass().getResource("").getPath();
+		//pathToClass = this.getClass().getResource("").getPath();
+		String path = MassSpec.class.getProtectionDomain().getCodeSource().getLocation().getPath();
+		try {
+			pathToClass = URLDecoder.decode(path, "UTF-8").replace("JAMSS.jar", "");
+		} catch (UnsupportedEncodingException ex) {
+			JOptionPane.showMessageDialog(null, "Error: encoding error when finding path to JAR file.", "Error", JOptionPane.ERROR_MESSAGE);
+		}
+		
 		// check for extant RT files and delete them
-		File directory = new File(pathToClass);
+		File directory = new File(pathToClass + File.separator + "JAMSSfiles");
+		directory.mkdir(); // create a new directory if doesn't exist
 		// get all .ser files
 		String[] myFiles = directory.list(new FilenameFilter() {
 			@Override
@@ -336,12 +346,14 @@ public class MassSpec {
 				return fileName.endsWith(".ser");
 			}
 		});
-		for (String fileName : myFiles){
-			// delete file
-			File file = new File(pathToClass + "/" + fileName);
-			file.delete();
+
+		if (myFiles != null){
+			for (String fileName : myFiles){
+				// delete file
+				File file = new File(pathToClass + File.separator + fileName);
+				file.delete();
+			}
 		}
-		
 		// Create thread-specific list of output
 		outputScans = new HashMap();
 		
@@ -375,7 +387,8 @@ public class MassSpec {
 		
 		// load the models
 		try {
-			rtCls = (Classifier) weka.core.SerializationHelper.read("/home/rob/Documents/mass_spec/mspireSimulator/src/simulatorGUI/"+rtModelLocation);
+			//rtCls = (Classifier) weka.core.SerializationHelper.read(pathToClass + "JAMSS.jar" + File.separator + "simulatorGUI" + File.separator + rtModelLocation);
+			rtCls = (Classifier) weka.core.SerializationHelper.read(this.getClass().getResourceAsStream(rtModelLocation) );
 		} catch (Exception ex) {
 			JOptionPane.showMessageDialog(null, "Error: Weka RT model could not be loaded.", "Error", JOptionPane.ERROR_MESSAGE);
             return;
@@ -1260,7 +1273,7 @@ public class MassSpec {
 	
 	public static boolean outputPre(int totalScans){
 		String filename = "output.mzML";
-		File outputFile = new File(filename);
+		File outputFile = new File(File.separator + "JAMSSfiles" + File.separator + filename);
 		try{
 			outputFile.delete(); // overwrite any previous truth file
 			outputFile.createNewFile(); 
@@ -1269,7 +1282,7 @@ public class MassSpec {
 			return false;
 		}
 		try {
-			FileWriter outputWriter = new FileWriter(filename, true);
+			FileWriter outputWriter = new FileWriter(File.separator + "JAMSSfiles" + File.separator + filename, true);
 			outputWriter.write("<?xml version=\"1.0\" encoding=\"UTF-8\"?>" + System.getProperty("line.separator"));
 			outputWriter.write("\t<mzML xmlns=\"http://psi.hupo.org/ms/mzml\" xmlns:xsi=\"http://www.w3.org/2001/XMLSchema-instance\" xmlns:xsd=\"http://www.w3.org/2001/XMLSchema\" xsi:schemaLocation=\"http://psi.hupo.org/ms/mzml http://psidev.info/files/ms/mzML/xsd/mzML1.1.0.xsd\" version=\"1.1.0\" id=\"ms1_and_ms2\">"+ System.getProperty("line.separator"));
 			outputWriter.write("\t\t<cvList count=\"3\">"+ System.getProperty("line.separator"));
@@ -1313,7 +1326,7 @@ public class MassSpec {
 	
 	public static boolean outputPost(){
 		try {
-			FileWriter mzMLWriter = new FileWriter("output.mzML",true); // append
+			FileWriter mzMLWriter = new FileWriter(File.separator + "JAMSSfiles" + File.separator + "output.mzML",true); // append
 
 			// write tail
 			mzMLWriter.write("\t\t</spectrumList>"+ System.getProperty("line.separator"));
@@ -1329,7 +1342,7 @@ public class MassSpec {
 	// output mzml file
 	public static int output(LinkedList<Centroid> masterScan, double rt, int scanIdx){
 		try {
-			FileWriter mzMLWriter = new FileWriter("output.mzML",true); // append
+			FileWriter mzMLWriter = new FileWriter(File.separator + "JAMSSfiles" + File.separator + "output.mzML",true); // append
 			int precursorIdx;
 			
 			String encodedMZ = compressCentroids(masterScan,true);
@@ -1480,7 +1493,7 @@ public class MassSpec {
 		}
 	}
 	static boolean printTruthPre(){
-		File peptideFile = new File("output_truth_peptides.csv");
+		File peptideFile = new File(File.separator + "JAMSSfiles" + File.separator + "output_truth_peptides.csv");
 		try{
 			peptideFile.createNewFile(); // overwrite any previous truth file
 		} catch(IOException e){
@@ -1488,7 +1501,7 @@ public class MassSpec {
 			return false;
 		}
 		
-		File truthFile = new File("output_truth.csv");
+		File truthFile = new File(File.separator + "JAMSSfiles" + File.separator + "output_truth.csv");
 		try{
 			truthFile.createNewFile(); // overwrite any previous truth file
 		} catch(IOException e){
@@ -1502,7 +1515,7 @@ public class MassSpec {
 		// print peptide file 
 		try {
 			//contains list of peptide sequences and IDs
-			FileWriter peptideWriter = new FileWriter("output_truth_peptides.csv",true); // append
+			FileWriter peptideWriter = new FileWriter(File.separator + "JAMSSfiles" + File.separator + "output_truth_peptides.csv",true); // append
 
 			for(int pepIdx = 0; pepIdx < peptides.size(); pepIdx++){
 				//proteinID (pos in fasta file), peptideID, peptide sequence
@@ -1517,7 +1530,7 @@ public class MassSpec {
 		// print regular truth file
 		try {
 			// centroidID, ionFeatureID, charge, pepID, proteinID, m/z, RT, intensity
-			FileWriter truthWriter = new FileWriter("output_truth.csv",true); // append
+			FileWriter truthWriter = new FileWriter(File.separator + "JAMSSfiles" + File.separator + "output_truth.csv",true); // append
 			for (Centroid cent : masterScan){
 				truthWriter.write(cent.centroidID + "," + cent.ionFeatureID + "," + cent.charge +"," + cent.pepID + "," + cent.proteinID + "," + cent.mz + "," + rt + "," + cent.abundance);
 			}
@@ -1536,7 +1549,7 @@ public class MassSpec {
 		for (int i=keys.length-1; i>=0; i--){
 			try{
 				// create a filename with the RT followed by time to distinguish it from other files at this RT
-				FileOutputStream fileOut = new FileOutputStream(pathToClass + "/" + keys[i] + "_" + msIdx + ".ser",true);
+				FileOutputStream fileOut = new FileOutputStream(pathToClass + File.separator + "JAMSSfiles" + File.separator + keys[i] + "_" + msIdx + ".ser",true);
 				ObjectOutputStream out = new ObjectOutputStream(fileOut);
 				out.writeObject((LinkedList<Centroid>)outputScans.get(keys[i]));
 				out.close();
@@ -1561,7 +1574,7 @@ public class MassSpec {
 		for(Double rt : rtList){
 			LinkedList<Centroid> masterScan = new LinkedList<Centroid>();
 			
-			File directory = new File(pathToClass);
+			File directory = new File(pathToClass + File.separator + "JAMSSfiles");
 
 			// get all .ser files
 			String[] myFiles = directory.list(new FilenameFilter() {
@@ -1572,18 +1585,18 @@ public class MassSpec {
 			});
 			
 			for (String fileName : myFiles){
-				if (fileName.replace(pathToClass,"").split("_")[0].equals(rt.toString())){
+				if (fileName.replace(pathToClass + File.separator + "JAMSSfiles" + File.separator,"").split("_")[0].equals(rt.toString())){
 					// open file and deserialize objects in file
 					LinkedList<Centroid> thisScan = null;
 					try
 					{
-					   FileInputStream fileIn = new FileInputStream(pathToClass + "/" + fileName);
+					   FileInputStream fileIn = new FileInputStream(pathToClass + File.separator + "JAMSSfiles" + File.separator + fileName);
 					   ObjectInputStream in = new ObjectInputStream(fileIn);
 					   thisScan = (LinkedList<Centroid>) in.readObject();
 					   in.close();
 					   fileIn.close();
 					   // delete file
-					   File file = new File(pathToClass + "/" + fileName);
+					   File file = new File(pathToClass + File.separator + "JAMSSfiles" + File.separator + fileName);
 					   file.delete();
 					}catch(IOException i)
 					{
